item type in hook itself
